Extension considered signed in when personal address is saved
Previously it was only considered signed in when both a personal and
private address were available. However, it's possible that the
extension won't be able to generate a private duck address (e.g. invalid
token, firewall, API down, etc.) This updates the logic to consider the
extension to be signed in when just a personal address is saved.