Position tooltip above input when no space below (#282)
* Position tooltip above input when no space below

* Preload webfonts before showing tooltip

This prevents the fonts loading in once the tooltip is shown and causing
a layout shift. This is more important when the tooltip is showing above
the input, as it can cause the tooltip to jump up.

* Check tooltip position after revealed and bail if called before

When running on slower devices, the tooltip position may be checked
before the tooltip is revealed. When this happens, we now bail early
without updating the position. Check position is now called after styles
have loaded, when we know the tooltip is visible.

* Hide caret initially and reduce position calculation calls

* Read hidden attribute from DOM

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>

* Fix linting issue with playwright report

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>

* Rename typoed file

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>

---------

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>
Co-authored-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>