passwords: require a digit in passwords generated from default rules (#562)
We previously required only a character in `[-!?$&#%]` when generating
passwords for sites with neither a `passwordrules` attribute on their
password input nor an override in `rules.json`. In many cases we'd
include a number by random chance, but the relative weight of digits
in the default pool meant they would sometimes not appear at all.
Since sites with password content rules commonly require a digit to be
present, this led to generated and saved passwords that immediately fail
password validation. Require a digit for passwords generated from the
default ruleset.