forms: ignore inputs without a width or height
We occasionally abandon valid forms that contain a large number of
`<input/>`s hidden within a `display: none` element (or some similar such
configuration). These elements can't be filled, and thus shouldn't be
classified immediately. Use an element's `boundingClientRect` to provide
a simple visibility heuristic (similar to `checkVisibility`[^1]), and use
it to ignore inputs that have a zero width or height. When those
elements change to a visible state later, they'll be reevaluated and
classified in a separate pass.