Update release workflow to support Apple monorepo (#756)
Asana: https://app.asana.com/0/72649045549333/1209197424401946

Description

This PR updates the release workflow to support the Apple monorepo.

There are a few TODOs left deliberately in place, which I'll resolve after the review. They're in place to allow the reviewer to test the workflow manually if desired, without causing noise for other platforms.